@@ -2,11 +2,11 @@
 	"name": "fuzeworks/core",
 	"description": "FuzeWorks Framework Core",
 	"homepage": "https://techfuze.net/fuzeworks",
-	"license": ["GPL-3.0"],
+	"license": ["MIT"],
 	"authors": [
 		{
-			"name": "Abel Hoogeveen",
-			"homepage": "https://myfuze.net"
+			"name": "TechFuze",
+			"homepage": "https://techfuze.net"
 		},
 		{
 			"name": "FuzeWorks Community",
@@ -14,21 +14,12 @@
 		}
 	],
 	"require": {
-		"php": ">=7.0.0",
-	    "ext-curl": "*",
-	    "ext-json": "*"
-	},
-	"suggest": {
-	    "smarty/smarty": "Allows using Smarty in templates",
-		"latte/latte": "Allows using Latte in templates",
-		"tracy/tracy": "Allows for extensive debugging"
+		"php": ">=7.1.0",
+		"psr/log": "1.1.0"
 	},
 	"require-dev": {
-        "phpunit/phpunit": "6.2.*",
-		"mikey179/vfsStream": "1.1.*",
-	    "smarty/smarty": "3.1.*",
-		"latte/latte": "2.4.*",
-		"tracy/tracy": "2.4.*"
+        "phpunit/phpunit": "^7",
+		"mikey179/vfsstream": "1.1.*"
 	},
 	"autoload": {
 		"psr-4": {
