@@ -1,71 +1,70 @@
 <?php
 /**
- * FuzeWorks.
+ * FuzeWorks Framework Core.
  *
- * The FuzeWorks MVC PHP FrameWork
+ * The FuzeWorks PHP FrameWork
  *
- * Copyright (C) 2015   TechFuze
+ * Copyright (C) 2013-2019 TechFuze
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  *
  * @author    TechFuze
- * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
- * @copyright Copyright (c) 1996 - 2015, Free Software Foundation, Inc. (http://www.fsf.org/)
- * @license   http://opensource.org/licenses/GPL-3.0 GPLv3 License
+ * @copyright Copyright (c) 2013 - 2019, TechFuze. (http://techfuze.net)
+ * @license   https://opensource.org/licenses/MIT MIT License
  *
  * @link  http://techfuze.net/fuzeworks
  * @since Version 0.0.1
  *
- * @version Version 1.0.0
+ * @version Version 1.2.0
  */
 
 namespace FuzeWorks;
 use FuzeWorks\ConfigORM\ConfigORM;
+use FuzeWorks\Event\ConfigGetEvent;
 use FuzeWorks\Exception\ConfigException;
+use FuzeWorks\Exception\EventException;
 
 /**
  * Config Class.
  *
  * This class gives access to the config files. It allows you to open configurations and edit them.
  *
- * @author    Abel Hoogeveen <abel@techfuze.net>
- * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
- * @todo      Implement config extensions
+ * @author    TechFuze <contact@techfuze.net>
+ * @copyright Copyright (c) 2013 - 2019, TechFuze. (http://techfuze.net)
  */
 class Config
 {
+    use ComponentPathsTrait;
 
     /**
      * Array where all config files are saved while FuzeWorks runs
      * 
      * @var array Array of all loaded config file ORM's
      */
-    protected $cfg = array();
+    protected $cfg = [];
 
     /**
-     * Paths where Helpers can be found. 
-     * 
-     * Libraries will only be loaded if either a directory is supplied or it is in one of the helperPaths
-     * 
-     * @var array Array of paths where helpers can be found
+     * Array of config values that will be overridden
+     *
+     * @var array of config values
      */
-    protected $configPaths = array();
-
-    public function __construct()
-    {
-        $this->configPaths[] = Core::$appDir . DS. 'Config';
-    }
+    public static $configOverrides = [];
 
     /**
      * Retrieve a config file object
@@ -75,30 +74,44 @@ class Config
      * @return  ConfigORM of the config file. Allows for easy reading and editing of the file
      * @throws  ConfigException
      */
-    public function getConfig($configName, array $configPaths = array()): ConfigORM
+    public function getConfig(string $configName, array $configPaths = []): ConfigORM
     {
-        // First determine what directories to use
-        $directories = (empty($configPaths) ? $this->configPaths : $configPaths);
-
         // Determine the config name
         $configName = strtolower($configName);
-        
+
         // If it's already loaded, return the existing object
         if (isset($this->cfg[$configName]))
         {
             return $this->cfg[$configName];
         }
 
+        // First determine what directories to use
+        $paths = [];
+        if (!empty($configPaths))
+            $paths[3] = $configPaths;
+        else
+            $paths = $this->componentPaths;
+
         // Otherwise try and load a new one
-        $this->cfg[$configName] = $this->loadConfigFile($configName, $directories);
+        $this->cfg[$configName] = $this->loadConfigFile($configName, $paths);
         return $this->cfg[$configName];
     }
-    
+
+    /**
+     * @param $configName
+     * @return ConfigORM
+     * @throws ConfigException
+     */
     public function get($configName): ConfigORM
     {
         return $this->getConfig($configName);
     }
 
+    /**
+     * @param $configName
+     * @return ConfigORM
+     * @throws ConfigException
+     */
     public function __get($configName): ConfigORM
     {
         return $this->getConfig($configName);
@@ -109,7 +122,7 @@ class Config
      */
     public function discardConfigFiles()
     {
-        $this->cfg = array();
+        $this->cfg = [];
     }
 
     /**
@@ -120,68 +133,92 @@ class Config
      * @return  ConfigORM of the config file. Allows for easy reading and editing of the file
      * @throws  ConfigException
      */
-    protected function loadConfigFile($configName, array $configPaths): ConfigORM
+    protected function loadConfigFile(string $configName, array $configPaths): ConfigORM
     {
-        // Cycle through all directories
-        foreach ($configPaths as $directory) 
+        // Fire event to intercept the loading of a config file
+        /** @var ConfigGetEvent $event */
+        try {
+            $event = Events::fireEvent('configGetEvent', $configName, $configPaths);
+            // @codeCoverageIgnoreStart
+        } catch (EventException $e) {
+            throw new ConfigException("Could not load config. ConfigGetEvent fired exception: '" . $e->getMessage() . "''", 1);
+            // @codeCoverageIgnoreEnd
+        }
+
+        // If cancelled, load empty config
+        if ($event->isCancelled())
+            return new ConfigORM();
+
+        // Cycle through all priorities if they exist
+        for ($i=Priority::getHighestPriority(); $i<=Priority::getLowestPriority(); $i++)
         {
-            // If file exists, load it and break the loop
-            $file = $directory . DS . 'config.'.$configName.'.php';
-            if (file_exists($file))
+            if (!isset($event->configPaths[$i]))
+                continue;
+
+            // Cycle through all directories
+            foreach ($event->configPaths[$i] as $configPath)
             {
-                // Load object
-                return new ConfigORM($file);
-                break;
+                // If file exists, load it and break the loop
+                $file = $configPath . DS . 'config.'.strtolower($event->configName).'.php';
+                if (file_exists($file))
+                {
+                    // Load object
+                    $configORM = (new ConfigORM())->load($file);
+
+                    // Override config values if they exist
+                    if (isset(self::$configOverrides[$event->configName]))
+                    {
+                        foreach (self::$configOverrides[$event->configName] as $configKey => $configValue)
+                            $configORM->{$configKey} = $configValue;
+                    }
+
+                    // Return object
+                    return $configORM;
+                }
             }
         }
 
         // Try fallback
-        $file = Core::$coreDir . DS . 'Config' . DS . 'config.' . $configName . '.php';
+        $file = Core::$coreDir . DS . 'Config' . DS . 'config.' . $event->configName . '.php';
         if (file_exists($file))
         {
             // Load object
-            return new ConfigORM($file);
+            $configORM = (new ConfigORM())->load($file);
+
+            // Override config values if they exist
+            if (isset(self::$configOverrides[$event->configName]))
+            {
+                foreach (self::$configOverrides[$event->configName] as $configKey => $configValue)
+                    $configORM->{$configKey} = $configValue;
+            }
+
+            // Return object
+            return $configORM;
         }
 
-        throw new ConfigException("Could not load config. File $configName not found", 1);
+        throw new ConfigException("Could not load config. File $event->configName not found", 1);
     }
 
     /**
-     * Add a path where config files can be found
-     * 
-     * @param string $directory The directory
-     * @return void
+     * Override a config value before FuzeWorks is loaded.
+     *
+     * Allows the user to change any value in config files loaded by FuzeWorks.
+     *
+     * @param string $configName
+     * @param string $configKey
+     * @param $configValue
      */
-    public function addConfigPath($directory)
+    public static function overrideConfig(string $configName, string $configKey, $configValue)
     {
-        if (!in_array($directory, $this->configPaths))
-        {
-            $this->configPaths[] = $directory;
-        }
-    }
+        // Convert configName
+        $configName = strtolower($configName);
 
-    /**
-     * Remove a path where config files can be found
-     * 
-     * @param string $directory The directory
-     * @return void
-     */    
-    public function removeConfigPath($directory)
-    {
-        if (($key = array_search($directory, $this->configPaths)) !== false) 
-        {
-            unset($this->configPaths[$key]);
-        }
-    }
+        // If config doesn't exist yet, create it
+        if (!isset(self::$configOverrides[$configName]))
+            self::$configOverrides[$configName] = [];
 
-    /**
-     * Get a list of all current configPaths
-     * 
-     * @return array Array of paths where config files can be found
-     */
-    public function getConfigPaths(): array
-    {
-        return $this->configPaths;
+        // And set the value
+        self::$configOverrides[$configName][$configKey] = $configValue;
     }
 
 }
