@@ -1,36 +1,42 @@
 <?php
 /**
- * FuzeWorks.
+ * FuzeWorks Framework Core.
  *
- * The FuzeWorks MVC PHP FrameWork
+ * The FuzeWorks PHP FrameWork
  *
- * Copyright (C) 2015   TechFuze
+ * Copyright (C) 2013-2019 TechFuze
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  *
  * @author    TechFuze
- * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
- * @copyright Copyright (c) 1996 - 2015, Free Software Foundation, Inc. (http://www.fsf.org/)
- * @license   http://opensource.org/licenses/GPL-3.0 GPLv3 License
+ * @copyright Copyright (c) 2013 - 2019, TechFuze. (http://techfuze.net)
+ * @license   https://opensource.org/licenses/MIT MIT License
  *
  * @link  http://techfuze.net/fuzeworks
  * @since Version 0.0.1
  *
- * @version Version 1.0.0
+ * @version Version 1.2.0
  */
 
 namespace FuzeWorks;
+use FuzeWorks\Event\HelperLoadEvent;
+use FuzeWorks\Exception\EventException;
 use FuzeWorks\Exception\HelperException;
 
 /**
@@ -49,53 +55,35 @@ use FuzeWorks\Exception\HelperException;
  * FuzeWorks does not load Helper Files by default, so the first step in using a Helper is to load it. Once loaded, 
  * it becomes globally available to everything. 
  *
- * @author    Abel Hoogeveen <abel@techfuze.net>
- * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
+ * @author    TechFuze <contact@techfuze.net>
+ * @copyright Copyright (c) 2013 - 2019, TechFuze. (http://techfuze.net)
  */
 class Helpers
 {
+    use ComponentPathsTrait;
 
     /**
      * Array of loadedHelpers, so that they won't be reloaded
      * 
      * @var array Array of loaded helperNames
      */
-    protected $helpers = array();
-
-    /**
-     * Paths where Helpers can be found. 
-     * 
-     * Libraries will only be loaded if either a directory is supplied or it is in one of the helperPaths
-     * 
-     * @var array Array of paths where helpers can be found
-     */
-    protected $helperPaths = array();
-
-    public function __construct()
-    {
-        $this->helperPaths = [
-            Core::$appDir . DS . 'Helpers',
-            Core::$coreDir . DS . 'Helpers'
-        ];
-    }
+    protected $helpers = [];
 
     /**
      * Load a helper.
-     * 
+     *
      * Supply the name and the helper will be loaded from the supplied directory,
      * or from one of the helperPaths (which you can add).
-     * 
-     * @param string        $helperName Name of the helper
-     * @param string|null   $directory  Directory to load the helper from, will ignore $helperPaths
-     * @return bool                     Whether the helper was succesfully loaded (true if yes)
+     *
+     * @param string $helperName Name of the helper
+     * @param array $helperPaths
+     * @return bool                     Whether the helper was successfully loaded (true if yes)
+     * @throws HelperException
      */
-    public function load($helperName, $directory = null): bool
+    public function load(string $helperName, array $helperPaths = []): bool
     {
-        // First determine the name of the helper
-        $helperName = strtolower(str_replace(array('_helper', '.php'), '', $helperName).'_helper');
-        
         // Determine what directories should be checked
-        $directories = (is_null($directory) ? $this->helperPaths : array($directory));
+        $helperPaths = (empty($helperPaths) ? $this->componentPaths : [3 => $helperPaths]);
 
         // Check it is already loaded
         if (isset($this->helpers[$helperName]))
@@ -104,62 +92,51 @@ class Helpers
             return false;
         }
 
-        // First check if there is an 'extension' class
-        $extendedHelper = Factory::getInstance()->config->get('main')->application_prefix . $helperName;
-        $extendedHelperLoaded = false;
-        foreach ($directories as $helperPath) 
-        {
-            $file = $helperPath . DS . $extendedHelper . '.php';
-            if (file_exists($file))
-            {
-                $extendedHelperLoaded = true;
-                $extendedHelperFile = $file;
-            }
+        /** @var HelperLoadEvent $event */
+        try {
+            $event = Events::fireEvent('helperLoadEvent', $helperName, $helperPaths);
+
+            // @codeCoverageIgnoreStart
+        } catch (EventException $e) {
+            throw new HelperException("Could not load helper. helperLoadEvent failed: '" . $e->getMessage() . "''");
+            // @codeCoverageIgnoreEnd
         }
 
-        // If an extension is loaded there needs to be a base helper
-        if ($extendedHelperLoaded)
+        // If cancelled by event, abort loading helper
+        if ($event->isCancelled())
         {
-            $baseHelper = Core::$coreDir . DS . 'Helpers' . DS . $helperName.'.php';
-            if (!file_exists($baseHelper))
-            {
-                throw new HelperException("Could not load helper. Base Helper not found while Extension loaded", 1);
-            }
-
-            // Fire the associated event
-            $event = Events::fireEvent('helperLoadEvent', $helperName, $baseHelper, $extendedHelper, $extendedHelperFile);
-            if ($event->isCancelled()) 
-            {
-                Logger::log("Not loading helper. Aborted by event");
-                return false;
-            }
-
-            include_once($event->extendedHelperFile);
-            include_once($event->helperFile);
-            $this->helpers[$event->helperName] = true;
-            Logger::log("Loading base helper '".$event->helperName."' and extended helper '".$event->extendedHelperName."'");
-            return true;
+            Logger::log("Not loading helper. Aborted by event");
+            return false;
         }
 
-        // If no extension exists, try loading a regular helper
-        foreach ($directories as $helperPath) 
+        // Iterate over helperPaths and attempt to load if helper exists
+        for ($i=Priority::getHighestPriority(); $i<=Priority::getLowestPriority(); $i++)
         {
-            $file = $helperPath . DS . $helperName . '.php';
-            if (file_exists($file))
-            {
+            if (!isset($event->helperPaths[$i]))
+                continue;
 
-                // Fire the associated event
-                $event = Events::fireEvent('helperLoadEvent', $helperName, $file);
-                if ($event->isCancelled()) 
+            foreach ($event->helperPaths[$i] as $helperPath)
+            {
+                $file = $helperPath . DS . $event->helperName . '.php';
+                $subfile = $helperPath . DS . $event->helperName . DS . $event->helperName . '.php';
+                if (file_exists($file))
                 {
-                    Logger::log("Not loading helper. Aborted by event");
-                    return false;
+                    // Load and register
+                    include_once($file);
+                    $this->helpers[$event->helperName] = true;
+                    Logger::log("Loaded helper '".$event->helperName."'");
+                    return true;
                 }
 
-                include_once($event->helperFile);
-                $this->helpers[$event->helperName] = true;
-                Logger::log("Loading helper '".$event->helperName."'");
-                return true;
+                // If php file not in main directory, check subdirectories
+                elseif (file_exists($subfile))
+                {
+                    // Load and register
+                    include_once($subfile);
+                    $this->helpers[$event->helperName] = true;
+                    Logger::log("Loaded helper '".$event->helperName."''");
+                    return true;
+                }
             }
         }
 
@@ -169,51 +146,14 @@ class Helpers
     /**
      * Alias for load
      * @see load() for more details
-     * 
-     * @param string        $helperName Name of the helper
-     * @param string|null   $directory  Directory to load the helper from, will ignore $helperPaths
-     * @return bool                     Whether the helper was succesfully loaded (true if yes)
+     *
+     * @param string $helperName Name of the helper
+     * @param array $helperPaths
+     * @return bool                     Whether the helper was successfully loaded (true if yes)
+     * @throws HelperException
      */
-    public function get($helperName, $directory = null): bool
+    public function get($helperName, array $helperPaths = []): bool
     {
-        return $this->load($helperName, $directory);
-    }
-
-    /**
-     * Add a path where helpers can be found
-     * 
-     * @param string $directory The directory
-     * @return void
-     */
-    public function addHelperPath($directory)
-    {
-        if (!in_array($directory, $this->helperPaths))
-        {
-            $this->helperPaths[] = $directory;
-        }
-    }
-
-    /**
-     * Remove a path where helpers can be found
-     * 
-     * @param string $directory The directory
-     * @return void
-     */    
-    public function removeHelperPath($directory)
-    {
-        if (($key = array_search($directory, $this->helperPaths)) !== false) 
-        {
-            unset($this->helperPaths[$key]);
-        }
-    }
-
-    /**
-     * Get a list of all current helperPaths
-     * 
-     * @return array Array of paths where helpers can be found
-     */
-    public function getHelperPaths(): array
-    {
-        return $this->helperPaths;
+        return $this->load($helperName, $helperPaths);
     }
 }
