@@ -1,33 +1,37 @@
 <?php
 /**
- * FuzeWorks.
+ * FuzeWorks Framework Core.
  *
- * The FuzeWorks MVC PHP FrameWork
+ * The FuzeWorks PHP FrameWork
  *
- * Copyright (C) 2015   TechFuze
+ * Copyright (C) 2013-2019 TechFuze
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  *
  * @author    TechFuze
- * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
- * @copyright Copyright (c) 1996 - 2015, Free Software Foundation, Inc. (http://www.fsf.org/)
- * @license   http://opensource.org/licenses/GPL-3.0 GPLv3 License
+ * @copyright Copyright (c) 2013 - 2019, TechFuze. (http://techfuze.net)
+ * @license   https://opensource.org/licenses/MIT MIT License
  *
  * @link  http://techfuze.net/fuzeworks
  * @since Version 0.0.1
  *
- * @version Version 1.0.0
+ * @version Version 1.2.0
  */
 
 namespace FuzeWorks\ConfigORM;
@@ -38,9 +42,8 @@ use FuzeWorks\Exception\ConfigException;
  *
  * Handles entries in the config directory of FuzeWorks and is able to dynamically update them when requested
  *
- * @author    Abel Hoogeveen <abel@techfuze.net>
- * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
- * @todo      Implement unit tests
+ * @author    TechFuze <contact@techfuze.net>
+ * @copyright Copyright (c) 2013 - 2019, TechFuze. (http://techfuze.net)
  */
 class ConfigORM extends ConfigORMAbstract
 {
@@ -52,28 +55,30 @@ class ConfigORM extends ConfigORMAbstract
     private $file;
 
     /**
-     * Sets up the class and the connection to the PHP file.
+     * Load the ConfigORM file.
      *
-     * @param string $filename The current filename
-     *
-     * @throws ConfigException on fatal error
+     * @param string $file
+     * @return ConfigORM
+     * @throws ConfigException
      */
-    public function __construct($file = null)
+    public function load(string $file = ''): ConfigORM
     {
-        if (is_null($file))
+        if (empty($file))
         {
             throw new ConfigException('Could not load config file. No file provided', 1);
         }
-        elseif (file_exists($file)) 
+        elseif (file_exists($file))
         {
             $this->file = $file;
             $this->cfg = (array) include $file;
             $this->originalCfg = $this->cfg;
-        } 
-        else 
+        }
+        else
         {
             throw new ConfigException('Could not load config file. Config file does not exist', 1);
         }
+
+        return $this;
     }
 
     /**
@@ -81,7 +86,7 @@ class ConfigORM extends ConfigORMAbstract
      *
      * @throws ConfigException on fatal error
      */
-    public function commit()
+    public function commit(): bool
     {
     	// Write the changes
         if (is_writable($this->file)) {
@@ -90,6 +95,6 @@ class ConfigORM extends ConfigORMAbstract
 
             return true;
         }
-        throw new ConfigException("Could not write config file. $file is not writable", 1);
+        throw new ConfigException("Could not write config file. $this->file is not writable", 1);
     }
 }
