@@ -1,33 +1,37 @@
 <?php
 /**
- * FuzeWorks.
+ * FuzeWorks Framework Core.
  *
- * The FuzeWorks MVC PHP FrameWork
+ * The FuzeWorks PHP FrameWork
  *
- * Copyright (C) 2015   TechFuze
+ * Copyright (C) 2013-2019 TechFuze
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  *
  * @author    TechFuze
- * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
- * @copyright Copyright (c) 1996 - 2015, Free Software Foundation, Inc. (http://www.fsf.org/)
- * @license   http://opensource.org/licenses/GPL-3.0 GPLv3 License
+ * @copyright Copyright (c) 2013 - 2019, TechFuze. (http://techfuze.net)
+ * @license   https://opensource.org/licenses/MIT MIT License
  *
  * @link  http://techfuze.net/fuzeworks
  * @since Version 0.0.1
  *
- * @version Version 1.0.0
+ * @version Version 1.2.0
  */
 
 namespace FuzeWorks\ConfigORM;
@@ -39,25 +43,24 @@ use Iterator;
  * This class implements the iterator, so a config file can be accessed using foreach.
  * A file can also be returned using toArray(), so it will be converted to an array
  *
- * @author    Abel Hoogeveen <abel@techfuze.net>
- * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
- * @todo      Implement unit tests
+ * @author    TechFuze <contact@techfuze.net>
+ * @copyright Copyright (c) 2013 - 2019, TechFuze. (http://techfuze.net)
  */
 abstract class ConfigORMAbstract implements Iterator
 {
     /**
      * The original state of a config file. Can be reverted to using revert().
      *
-     * @var StdObject Config file
+     * @var array Config file
      */
-    protected $originalCfg;
+    protected $originalCfg = [];
 
     /**
      * The current state of a config file.
      *
-     * @var StdObject Config file
+     * @var array Config file
      */
-    protected $cfg;
+    protected $cfg = [];
 
     /**
      * Revert to the original conditions of the config file.
@@ -68,11 +71,32 @@ abstract class ConfigORMAbstract implements Iterator
     }
 
     /**
+     * Clears the current config
+     */
+    public function clear()
+    {
+        $this->cfg = [];
+    }
+
+    /**
+     * Replaces keys and values in the config with the provided array
+     *
+     * @param array $replacementArray
+     */
+    public function replace(array $replacementArray)
+    {
+        foreach ($replacementArray as $key => $value)
+        {
+            $this->cfg[$key] = $value;
+        }
+    }
+
+    /**
      * Checks if a requested key is set in the config file.
      *
      * @param string $name Parameter name
-     *
      * @return bool true on isset, false on not
+     * @codeCoverageIgnore
      */
     public function __isset($name)
     {
@@ -83,10 +107,22 @@ abstract class ConfigORMAbstract implements Iterator
      * Return a value from a config file.
      *
      * @param string $name Key of the requested entry
-     *
      * @return mixed Value of the requested entry
+     * @codeCoverageIgnore
      */
     public function __get($name)
+    {
+        return $this->cfg[$name];
+    }
+
+    /**
+     * Return a value from a config file.
+     *
+     * @param string $name Key of the requested entry
+     * @return mixed Value of the requested entry
+     * @codeCoverageIgnore
+     */
+    public function get($name)
     {
         return $this->cfg[$name];
     }
@@ -96,8 +132,21 @@ abstract class ConfigORMAbstract implements Iterator
      *
      * @param string $name  Key of the entry
      * @param mixed  $value Value of the entry
+     * @codeCoverageIgnore
      */
     public function __set($name, $value)
+    {
+        $this->cfg[$name] = $value;
+    }
+
+    /**
+     * Sets an entry in the config file.
+     *
+     * @param string $name  Key of the entry
+     * @param mixed  $value Value of the entry
+     * @codeCoverageIgnore
+     */
+    public function set($name, $value)
     {
         $this->cfg[$name] = $value;
     }
@@ -106,6 +155,7 @@ abstract class ConfigORMAbstract implements Iterator
      * Unset a value in a config file.
      *
      * @param string Key of the entry
+     * @codeCoverageIgnore
      */
     public function __unset($name)
     {
@@ -114,6 +164,7 @@ abstract class ConfigORMAbstract implements Iterator
 
     /**
      * Iterator method.
+     * @codeCoverageIgnore
      */
     public function rewind()
     {
@@ -122,6 +173,7 @@ abstract class ConfigORMAbstract implements Iterator
 
     /**
      * Iterator method.
+     * @codeCoverageIgnore
      */
     public function current()
     {
@@ -130,6 +182,7 @@ abstract class ConfigORMAbstract implements Iterator
 
     /**
      * Iterator method.
+     * @codeCoverageIgnore
      */
     public function key()
     {
@@ -138,6 +191,7 @@ abstract class ConfigORMAbstract implements Iterator
 
     /**
      * Iterator method.
+     * @codeCoverageIgnore
      */
     public function next()
     {
@@ -146,6 +200,7 @@ abstract class ConfigORMAbstract implements Iterator
 
     /**
      * Iterator method.
+     * @codeCoverageIgnore
      */
     public function valid()
     {
