@@ -1,46 +1,52 @@
 <?php
 /**
- * FuzeWorks.
+ * FuzeWorks Framework Core.
  *
- * The FuzeWorks MVC PHP FrameWork
+ * The FuzeWorks PHP FrameWork
  *
- * Copyright (C) 2015   TechFuze
+ * Copyright (C) 2013-2019 TechFuze
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  *
  * @author    TechFuze
- * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
- * @copyright Copyright (c) 1996 - 2015, Free Software Foundation, Inc. (http://www.fsf.org/)
- * @license   http://opensource.org/licenses/GPL-3.0 GPLv3 License
+ * @copyright Copyright (c) 2013 - 2019, TechFuze. (http://techfuze.net)
+ * @license   https://opensource.org/licenses/MIT MIT License
  *
  * @link  http://techfuze.net/fuzeworks
  * @since Version 0.0.1
  *
- * @version Version 1.0.0
+ * @version Version 1.2.0
  */
 
 namespace FuzeWorks;
-use FuzeWorks\Exception\Exception;
+
 use FuzeWorks\Exception\CoreException;
+use FuzeWorks\Exception\EventException;
 
 /**
  * FuzeWorks Core.
  *
  * Holds all the modules and starts the framework. Allows for starting and managing modules
  *
- * @author    Abel Hoogeveen <abel@techfuze.net>
- * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
+ * @todo Implement unit tests for autoloader() methods
+ * @author    TechFuze <contact@techfuze.net>
+ * @copyright Copyright (c) 2013 - 2019, TechFuze. (http://techfuze.net)
  */
 class Core
 {
@@ -49,7 +55,7 @@ class Core
      *
      * @var string Framework version
      */
-    public static $version = '1.0.0';
+    public static $version = '1.2.0';
 
     /**
      * Working directory of the Framework.
@@ -60,10 +66,6 @@ class Core
      */
     public static $cwd;
 
-    public static $appDir;
-
-    public static $wwwDir;
-
     public static $coreDir;
 
     public static $tempDir;
@@ -71,30 +73,38 @@ class Core
     public static $logDir;
 
     /**
-     * The HTTP status code of the current request
+     * Array of exception handlers, sorted by priority
      *
-     * @var int $http_status_code Status code
+     * @var array
      */
-    public static $http_status_code = 200;
+    protected static $exceptionHandlers = [];
+
+    /**
+     * Array of error handlers, sorted by priority
+     *
+     * @var array
+     */
+    protected static $errorHandlers = [];
+
+    /**
+     * Array of all classMaps which can be autoloaded.
+     *
+     * @var array
+     */
+    protected static $autoloadMap = [];
 
     /**
      * Initializes the core.
      *
      * @throws \Exception
      */
-    public static function init()
+    public static function init(): Factory
     {
         // Set the CWD for usage in the shutdown function
         self::$cwd = getcwd();
 
         // Set the core dir for when the loading of classes is required
         self::$coreDir = dirname(__DIR__);
-
-        // If the environment is not yet defined, use production settings
-        if (!defined('ENVIRONMENT'))
-        {
-            define('ENVIRONMENT', 'PRODUCTION');
-        }
         
         // Defines the time the framework starts. Used for timing functions in the framework
         if (!defined('STARTTIME')) {
@@ -105,53 +115,37 @@ class Core
         // Load basics
         ignore_user_abort(true);
         register_shutdown_function(array('\FuzeWorks\Core', 'shutdown'));
+        set_error_handler(array('\FuzeWorks\Core', 'errorHandler'), E_ALL);
+        set_exception_handler(array('\FuzeWorks\Core', 'exceptionHandler'));
+        spl_autoload_register(['\FuzeWorks\Core', 'autoloader'], true,false);
 
-        // Load core functionality
-        $container = new Factory();
-
-        // Load the config file of the FuzeWorks core
-        $config = $container->config->get('core');
-
-        // Disable events if requested to do so
-        if (!$config->enable_events)
-        {
-            Events::disable();
-        }
-
-        // And initialize multiple classes
-        $container->layout->init();
-        Language::init();
-
-        // And load all the plugins
-        $container->plugins->loadHeaders();
-
-        // And fire the coreStartEvent
-        $event = Events::fireEvent('coreStartEvent');
-        if ($event->isCancelled()) {
-            exit;
-        }
+        // Return the Factory
+        return new Factory();
     }
 
     /**
      * Stop FuzeWorks and run all shutdown functions.
      *
      * Afterwards run the Logger shutdown function in order to possibly display the log
+     * @throws EventException
      */
     public static function shutdown()
     {
         // Fix Apache bug where CWD is changed upon shutdown
         chdir(self::$cwd);
 
+        // Log the shutdown
+        Logger::newLevel("Shutting FuzeWorks down gracefully");
+
         // Fire the Shutdown event
         $event = Events::fireEvent('coreShutdownEvent');
-
         if ($event->isCancelled() === false)
         {
-            // If the output should be displayed, send the final render and parse the logger
             Logger::shutdownError();
-            Factory::getInstance()->output->_display();
             Logger::shutdown();
         }
+
+        Logger::stopLevel();
     }
 
     /**
@@ -173,35 +167,140 @@ class Core
         return $_is_php[$version];
     }
 
-    public static function isCli(): bool
+    public static function exceptionHandler()
     {
-        return (PHP_SAPI === 'cli' OR defined('STDIN'));
+        for ($i = Priority::getHighestPriority(); $i <= Priority::getLowestPriority(); $i++)
+        {
+            if (!isset(self::$exceptionHandlers[$i]))
+                continue;
+
+            foreach (self::$exceptionHandlers[$i] as $handler)
+                call_user_func_array($handler, func_get_args());
+        }
+    }
+
+    public static function errorHandler()
+    {
+        for ($i = Priority::getHighestPriority(); $i <= Priority::getLowestPriority(); $i++)
+        {
+            if (!isset(self::$errorHandlers[$i]))
+                continue;
+
+            foreach (self::$errorHandlers[$i] as $handler)
+                call_user_func_array($handler, func_get_args());
+        }
     }
 
     /**
-     * Is HTTPS?
+     * Add an exception handler to be called when an exception occurs
      *
-     * Determines if the application is accessed via an encrypted
-     * (HTTPS) connection.
-     *
-     * @return  bool
+     * @param callable $callback
+     * @param int $priority
      */
-    public static function isHttps(): bool
+    public static function addExceptionHandler(callable $callback, int $priority = Priority::NORMAL)
     {
-        if ( ! empty($_SERVER['HTTPS']) && strtolower($_SERVER['HTTPS']) !== 'off')
-        {
-            return TRUE;
-        }
-        elseif (isset($_SERVER['HTTP_X_FORWARDED_PROTO']) && $_SERVER['HTTP_X_FORWARDED_PROTO'] === 'https')
-        {
-            return TRUE;
-        }
-        elseif ( ! empty($_SERVER['HTTP_FRONT_END_HTTPS']) && strtolower($_SERVER['HTTP_FRONT_END_HTTPS']) !== 'off')
-        {
-            return TRUE;
-        }
+        if (!isset(self::$exceptionHandlers[$priority]))
+            self::$exceptionHandlers[$priority] = [];
 
-        return FALSE;
+        if (!in_array($callback, self::$exceptionHandlers[$priority]))
+            self::$exceptionHandlers[$priority][] = $callback;
+    }
+
+    /**
+     * Remove an exception handler from the list
+     *
+     * @param callable $callback
+     * @param int $priority
+     */
+    public static function removeExceptionHandler(callable $callback, int $priority = Priority::NORMAL)
+    {
+        if (isset(self::$exceptionHandlers[$priority]) && in_array($callback, self::$exceptionHandlers[$priority]))
+        {
+            foreach (self::$exceptionHandlers[$priority] as $i => $_callback)
+                if ($callback == $_callback)
+                    unset(self::$exceptionHandlers[$priority][$i]);
+        }
+    }
+
+    /**
+     * Add an error handler to be called when an error occurs
+     *
+     * @param callable $callback
+     * @param int $priority
+     */
+    public static function addErrorHandler(callable $callback, int $priority = Priority::NORMAL)
+    {
+        if (!isset(self::$errorHandlers[$priority]))
+            self::$errorHandlers[$priority] = [];
+
+        if (!in_array($callback, self::$errorHandlers[$priority]))
+            self::$errorHandlers[$priority][] = $callback;
+    }
+
+    /**
+     * Remove an error handler from the list
+     *
+     * @param callable $callback
+     * @param int $priority
+     */
+    public static function removeErrorHandler(callable $callback, int $priority = Priority::NORMAL)
+    {
+        if (isset(self::$errorHandlers[$priority]) && in_array($callback, self::$errorHandlers[$priority]))
+        {
+            foreach (self::$errorHandlers[$priority] as $i => $_callback)
+                if ($callback == $_callback)
+                    unset(self::$errorHandlers[$priority][$i]);
+        }
+    }
+
+    /**
+     * @param string $nameSpacePrefix
+     * @param string $filePath
+     * @throws CoreException
+     */
+    public static function addAutoloadMap(string $nameSpacePrefix, string $filePath)
+    {
+        // Remove leading slashes
+        $nameSpacePrefix = ltrim($nameSpacePrefix, '\\');
+
+        if (isset(self::$autoloadMap[$nameSpacePrefix]))
+            throw new CoreException("Could not add classes to autoloader. ClassMap already exists.");
+
+        if (!file_exists($filePath) && !is_dir($filePath))
+            throw new CoreException("Could not add classes to autoloader. Provided filePath does not exist.");
+
+        self::$autoloadMap[$nameSpacePrefix] = $filePath;
+    }
+
+    public static function autoloader(string $class)
+    {
+        // Remove leading slashes
+        $class = ltrim($class, '\\');
+
+        // First attempt and find if the prefix of the class is in the autoloadMap
+        foreach (self::$autoloadMap as $prefix => $path)
+        {
+            // If not, try next
+            if (strpos($class, $prefix) === false)
+                continue;
+
+            // If it contains the prefix, attempt to find the file
+            $className = substr($class, strlen($prefix) + 1);
+            $filePath = $path . DIRECTORY_SEPARATOR . str_replace('\\', DIRECTORY_SEPARATOR, $className) . '.php';
+            if (file_exists($filePath) && is_file($filePath))
+                require($filePath);
+        }
+    }
+
+    /**
+     * Clears the autoloader to its original state.
+     *
+     * Not intended for use by developer. Only for use during testing
+     * @internal
+     */
+    public static function clearAutoloader()
+    {
+        self::$autoloadMap = [];
     }
 
     /**
@@ -249,93 +348,12 @@ class Core
     }
 
     /**
-     * Set HTTP Status Header
+     * Whether the current environment is a production environment
      *
-     * @param   int the status code
-     * @param   string
-     * @return  void
+     * @return bool
      */
-    public static function setStatusHeader($code = 200, $text = '')
+    public static function isProduction(): bool
     {
-        if (self::isCli())
-        {
-            return;
-        }
-
-        if (empty($code) OR ! is_numeric($code))
-        {
-            throw new Exception('Status codes must be numeric', 1);
-        }
-
-        if (empty($text))
-        {
-            is_int($code) OR $code = (int) $code;
-            $stati = array(
-                100 => 'Continue',
-                101 => 'Switching Protocols',
-
-                200 => 'OK',
-                201 => 'Created',
-                202 => 'Accepted',
-                203 => 'Non-Authoritative Information',
-                204 => 'No Content',
-                205 => 'Reset Content',
-                206 => 'Partial Content',
-
-                300 => 'Multiple Choices',
-                301 => 'Moved Permanently',
-                302 => 'Found',
-                303 => 'See Other',
-                304 => 'Not Modified',
-                305 => 'Use Proxy',
-                307 => 'Temporary Redirect',
-
-                400 => 'Bad Request',
-                401 => 'Unauthorized',
-                402 => 'Payment Required',
-                403 => 'Forbidden',
-                404 => 'Not Found',
-                405 => 'Method Not Allowed',
-                406 => 'Not Acceptable',
-                407 => 'Proxy Authentication Required',
-                408 => 'Request Timeout',
-                409 => 'Conflict',
-                410 => 'Gone',
-                411 => 'Length Required',
-                412 => 'Precondition Failed',
-                413 => 'Request Entity Too Large',
-                414 => 'Request-URI Too Long',
-                415 => 'Unsupported Media Type',
-                416 => 'Requested Range Not Satisfiable',
-                417 => 'Expectation Failed',
-                422 => 'Unprocessable Entity',
-
-                500 => 'Internal Server Error',
-                501 => 'Not Implemented',
-                502 => 'Bad Gateway',
-                503 => 'Service Unavailable',
-                504 => 'Gateway Timeout',
-                505 => 'HTTP Version Not Supported'
-            );
-
-            if (isset($stati[$code]))
-            {
-                $text = $stati[$code];
-            }
-            else
-            {
-                throw new CoreException('No status text available. Please check your status code number or supply your own message text.', 1);
-            }
-        }
-
-        if (strpos(PHP_SAPI, 'cgi') === 0)
-        {
-            header('Status: '.$code.' '.$text, TRUE);
-        }
-        else
-        {
-            $server_protocol = isset($_SERVER['SERVER_PROTOCOL']) ? $_SERVER['SERVER_PROTOCOL'] : 'HTTP/1.1';
-            header($server_protocol.' '.$code.' '.$text, TRUE, $code);
-        }
+        return (ENVIRONMENT === 'PRODUCTION');
     }
 }
