@@ -1,36 +1,43 @@
 <?php
 /**
- * FuzeWorks.
+ * FuzeWorks Framework Core.
  *
- * The FuzeWorks MVC PHP FrameWork
+ * The FuzeWorks PHP FrameWork
  *
- * Copyright (C) 2015   TechFuze
+ * Copyright (C) 2013-2019 TechFuze
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  *
  * @author    TechFuze
- * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
- * @copyright Copyright (c) 1996 - 2015, Free Software Foundation, Inc. (http://www.fsf.org/)
- * @license   http://opensource.org/licenses/GPL-3.0 GPLv3 License
+ * @copyright Copyright (c) 2013 - 2019, TechFuze. (http://techfuze.net)
+ * @license   https://opensource.org/licenses/MIT MIT License
  *
  * @link  http://techfuze.net/fuzeworks
  * @since Version 0.0.1
  *
- * @version Version 1.0.0
+ * @version Version 1.2.0
  */
 
 namespace FuzeWorks;
+use FuzeWorks\Exception\ConfigException;
+use FuzeWorks\Exception\CoreException;
+use FuzeWorks\Exception\EventException;
 use FuzeWorks\Exception\FactoryException;
 
 /**
@@ -51,8 +58,8 @@ use FuzeWorks\Exception\FactoryException;
  * 
  * The Factory class is also extendible. This allows classes that extend Factory to access all it's properties. 
  * 
- * @author    Abel Hoogeveen <abel@techfuze.net>
- * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
+ * @author    TechFuze <contact@techfuze.net>
+ * @copyright Copyright (c) 2013 - 2019, TechFuze. (http://techfuze.net)
  */
 class Factory
 {
@@ -64,12 +71,12 @@ class Factory
 	 */
 	private static $sharedFactoryInstance;
 
-	/**
-	 * Whether to clone all Factory instances upon calling Factory::getInstance()
-	 * 
-	 * @var bool Clone all Factory instances.
-	 */
-	protected static $cloneInstances = false;
+    /**
+     * Whether the Factory has been initialized or not
+     *
+     * @var bool $initialized
+     */
+	private $initialized = false;
 
 	/**
 	 * Config Object
@@ -90,18 +97,6 @@ class Factory
 	public $events;
 	
 	/**
-	 * Models Object
-	 * @var Models
-	 */
-	public $models;
-	   
-	/**
-	 * Layout Object
-	 * @var Layout
-	 */
-	public $layout;
-	
-	/**
 	 * Libraries Object
 	 * @var Libraries
 	 */
@@ -114,63 +109,16 @@ class Factory
 	public $helpers;
 	
 	/**
-	 * Database Object
-	 * @var Database
-	 */
-	public $database;
-	
-	/**
-	 * Language Object
-	 * @var Language
-	 */
-	public $language;
-	
-	/**
-	 * Utf8 Object
-	 * @var Utf8
-	 */
-	public $utf8;
-	
-	/**
-	 * URI Object
-	 * @var URI
-	 */
-	public $uri;
-	
-	/**
-	 * Security Object
-	 * @var Security
-	 */
-	public $security;
-	
-	/**
-	 * Input Object
-	 * @var Input
-	 */
-	public $input;
-	
-	/**
-	 * Output Object
-	 * @var Output
-	 */
-	public $output;
-	
-	/**
-	 * Router Object
-	 * @var Router
-	 */
-	public $router;
-	
-	/**
 	 * Plugins Object
 	 * @var Plugins
 	 */
 	public $plugins;
 
-	/**
-	 * Factory instance constructor. Should only really be called once
-	 * @return void
-	 */
+    /**
+     * Factory instance constructor. Should only really be called once
+     * @throws ConfigException
+     * @throws FactoryException
+     */
 	public function __construct()
 	{
 		// If there is no sharedFactoryInstance, prepare it
@@ -181,21 +129,11 @@ class Factory
 	        $this->config = new Config();
 	        $this->logger = new Logger();
 	        $this->events = new Events();
-	        $this->models = new Models();
-	        $this->layout = new Layout();
 	        $this->libraries = new Libraries();
 	        $this->helpers = new Helpers();
-	        $this->database = new Database();
-	        $this->language = new Language();
-	        $this->utf8 = new Utf8();
-	        $this->uri = new URI();
-	        $this->output = new Output();
-	        $this->security = new Security();
-	        $this->input = new Input();
-	        $this->router = new Router();
 	        $this->plugins = new Plugins();
 
-	        return true;
+	        return;
 		}
 		// @codeCoverageIgnoreEnd
 
@@ -206,53 +144,85 @@ class Factory
 		    $this->{$key} = $value;
 		}
 
-		return true;
+		return;
 	}
 
-	/**
-	 * Get a new instance of the Factory class. 
-	 * 
-	 * @param bool $cloneInstance Whether to get a cloned instance (true) or exactly the same instance (false)
-	 * @return Factory Instance
-	 */
-	public static function getInstance($cloneInstance = false): Factory
-	{	
-		if ($cloneInstance === true || self::$cloneInstances === true)
-		{
-			return clone self::$sharedFactoryInstance;
-		}
+    /**
+     * Finalizes the Factory and sends out a coreStartEvent
+     *
+     * @return Factory
+     * @throws CoreException
+     */
+	public function initFactory(): Factory
+    {
+        // If already initialized, cancel
+        if ($this->initialized)
+            return $this;
 
-		return self::$sharedFactoryInstance;
+        // Load the config file of the FuzeWorks core
+        try {
+            $cfg = $this->config->get('core');
+        } catch (ConfigException $e) {
+            throw new CoreException("Could not initiate Factory. Config 'core' could not be found.");
+        }
+
+        // Disable events if requested to do so
+        if (!$cfg->get('enable_events'))
+            Events::disable();
+
+        // Initialize all components
+        foreach ($this as $component)
+        {
+            if (method_exists($component, 'init'))
+                $component->init();
+        }
+
+        // Initialize all plugins
+        $this->plugins->loadHeadersFromPluginPaths();
+
+        // Log actions
+        Logger::logInfo("FuzeWorks initialized. Firing coreStartEvent.");
+
+        // And fire the coreStartEvent
+        try {
+            Events::fireEvent('coreStartEvent');
+        } catch (EventException $e) {
+            throw new CoreException("Could not initiate Factory. coreStartEvent threw exception: ".$e->getMessage());
+        }
+
+        return $this;
+    }
+
+    /**
+     * Get an instance of a componentClass.
+     *
+     * @param string|null $instanceName
+     * @return mixed
+     * @throws FactoryException
+     */
+	public static function getInstance(string $instanceName = null)
+    {
+	    if (is_null($instanceName))
+	        return self::$sharedFactoryInstance;
+
+        // Determine the instance name
+        $instanceName = strtolower($instanceName);
+
+	    if (!isset(self::$sharedFactoryInstance->{$instanceName}))
+	        throw new FactoryException("Could not get instance. Instance was not found.");
+
+	    return self::$sharedFactoryInstance->{$instanceName};
 	}
 
-	/**
-	 * Enable cloning all Factory instances upon calling Factory::getInstance()
-	 * 
-	 * @return void
-	 */
-	public static function enableCloneInstances()
-	{
-		self::$cloneInstances = true;
-	}
-
-	/**
-	 * Disable cloning all Factory instances upon calling Factory::getInstance()
-	 * 
-	 * @return void
-	 */
-	public static function disableCloneInstances()
-	{
-		self::$cloneInstances = false;
-	}
-
-	/**
-	 * Create a new instance of one of the loaded classes.
-	 * It reloads the class. It does NOT clone it. 
-	 * 
-	 * @param string $className The name of the loaded class, WITHOUT the namespace
-	 * @param string $namespace Optional namespace. Defaults to 'FuzeWorks\'
-	 * @return Factory Instance
-	 */
+    /**
+     * Create a new instance of one of the loaded classes.
+     * It reloads the class. It does NOT clone it.
+     *
+     * @param string $className The name of the loaded class, WITHOUT the namespace
+     * @param string $namespace Optional namespace. Defaults to 'FuzeWorks\'
+     * @return Factory Instance
+     * @throws FactoryException
+     */
 	public function newInstance($className, $namespace = 'FuzeWorks\\'): self
 	{
 		// Determine the class to load
@@ -278,42 +248,47 @@ class Factory
 		return $this;
 	}
 
-	/**
-	 * Clone an instance of one of the loaded classes.
-	 * It clones the class. It does NOT re-create it. 
-	 * 
-	 * @param string $className The name of the loaded class, WITHOUT the namespace
-	 * @return Factory Instance
-	 */
-	public function cloneInstance($className): self
+    /**
+     * Clone an instance of one of the loaded classes.
+     * It clones the class. It does NOT re-create it.
+     *
+     * If the $onlyReturn = true is provided, the cloned instance will only be returned, and not set to the factory.
+     *
+     * @param string $className The name of the loaded class, WITHOUT the namespace
+     * @param bool $onlyReturn
+     * @return mixed
+     * @throws FactoryException
+     */
+	public static function cloneInstance(string $className, bool $onlyReturn = false)
 	{
 		// Determine the class to load
 		$instanceName = strtolower($className);
 
-		if (!isset($this->{$instanceName}))
-		{
+		if (!isset(self::$sharedFactoryInstance->{$instanceName}))
 			throw new FactoryException("Could not clone instance of '".$instanceName."'. Instance was not found.", 1);
-		}
+
+		if ($onlyReturn)
+		    return clone self::$sharedFactoryInstance->{$instanceName};
 
 		// Clone the instance
-		$this->{$instanceName} = clone $this->{$instanceName};
+		self::$sharedFactoryInstance->{$instanceName} = clone self::$sharedFactoryInstance->{$instanceName};
 
 		// Return itself
-		return $this;
+		return self::$sharedFactoryInstance->{$instanceName};
 	}
 
 	/**
 	 * Set an instance of one of the loaded classes with your own $object.
 	 * Replace the existing class with one of your own.
 	 * 
-	 * @param string $className The name of the loaded class, WITHOUT the namespace
+	 * @param string $objectName The name of the loaded class, WITHOUT the namespace
 	 * @param mixed  $object    Object to replace the class with
 	 * @return Factory Instance
 	 */
-	public function setInstance($className, $object): self
+	public function setInstance($objectName, $object): self
 	{
 		// Determine the instance name
-		$instanceName = strtolower($className);
+		$instanceName = strtolower($objectName);
 
 		// Unset and set
 		unset($this->{$instanceName});
@@ -323,12 +298,13 @@ class Factory
 		return $this;
 	}
 
-	/**
-	 * Remove an instance of one of the loaded classes. 
-	 * 
-	 * @param string $className The name of the loaded class, WITHOUT the namespace
-	 * @return Factory Factory Instance
-	 */
+    /**
+     * Remove an instance of one of the loaded classes.
+     *
+     * @param string $className The name of the loaded class, WITHOUT the namespace
+     * @return Factory Factory Instance
+     * @throws FactoryException
+     */
 	public function removeInstance($className): self
 	{
 		// Determine the instance name
@@ -345,4 +321,15 @@ class Factory
 		// Return itself
 		return $this;
 	}
+
+    /**
+     * Returns true if component is part of this Factory.
+     *
+     * @param $componentName
+     * @return bool
+     */
+    public function instanceIsset($componentName)
+    {
+        return isset($this->{$componentName});
+    }
 }
