@@ -1,4 +1,38 @@
 <?php
+/**
+ * FuzeWorks Framework Core.
+ *
+ * The FuzeWorks PHP FrameWork
+ *
+ * Copyright (C) 2013-2019 TechFuze
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ *
+ * @author    TechFuze
+ * @copyright Copyright (c) 2013 - 2019, TechFuze. (http://techfuze.net)
+ * @license   https://opensource.org/licenses/MIT MIT License
+ *
+ * @link  http://techfuze.net/fuzeworks
+ * @since Version 1.2.0
+ *
+ * @version Version 1.2.0
+ */
 if (!function_exists('getColoredString'))
 {
     function getColoredString($string, $foreground_color, $background_color) {
@@ -56,7 +90,7 @@ $id = 1;
 
 printf($mask, 'Id', 'Title', 'Runtime');
 printf($mask, $id, getColoredString('FuzeWorks debug log', 'black', 'light_gray'), '0 ms');
-foreach ($this->assigned_variables['Logs'] as $log) {
+foreach ($logs as $log) {
     $id++;
 
     $string = '';
@@ -68,6 +102,11 @@ foreach ($this->assigned_variables['Logs'] as $log) {
     elseif ($log['type'] == 'ERROR')
     {
         $string .= getColoredString('[ERROR]', 'black', 'red') . ' - ';
+        $string .= getColoredString($log['message'], 'black', 'red');
+    }
+    elseif ($log['type'] == 'EXCEPTION')
+    {
+        $string .= getColoredString('[EXCEPTION]', 'black', 'red') . ' - ';
         $string .= getColoredString($log['message'], 'black', 'red');
     }
     elseif ($log['type'] == "LEVEL_STOP")
